+++ src/Aggregation/SumBucket.php
@@ -2,8 +2,6 @@
 
 namespace Elastica\Aggregation;
 
-use Elastica\Exception\InvalidException;
-
 /**
  * Class SumBucket.
  *
@@ -14,17 +12,11 @@ class SumBucket extends AbstractAggregation implements GapPolicyInterface
     use Traits\BucketsPathTrait;
     use Traits\GapPolicyTrait;
 
-    public function __construct(string $name, ?string $bucketsPath = null)
+    public function __construct(string $name, string $bucketsPath)
     {
         parent::__construct($name);
 
-        if (null !== $bucketsPath) {
-            $this->setBucketsPath($bucketsPath);
-        } elseif (\func_num_args() >= 2) {
-            \trigger_deprecation('ruflin/elastica', '7.1.3', 'Passing null as 2nd argument to "%s()" is deprecated, pass a string instead. It will be removed in 8.0.', __METHOD__);
-        } else {
-            \trigger_deprecation('ruflin/elastica', '7.1.3', 'Not passing a 2nd argument to "%s()" is deprecated, pass a string instead. It will be removed in 8.0.', __METHOD__);
-        }
+        $this->setBucketsPath($bucketsPath);
     }
 
     /**
@@ -35,17 +27,5 @@ class SumBucket extends AbstractAggregation implements GapPolicyInterface
     public function setFormat(string $format): self
     {
         return $this->setParam('format', $format);
-    }
-
-    /**
-     * @throws InvalidException If buckets path or script is not set
-     */
-    public function toArray(): array
-    {
-        if (!$this->hasParam('buckets_path')) {
-            throw new InvalidException('Buckets path is required');
-        }
-
-        return parent::toArray();
     }
 }